spreading done
