+++ genomepy/logging.py
@@ -1,0 +1,14 @@
+import sys
+
+from loguru import logger
+
+# No traceback
+# sys.tracebacklimit = 0
+
+# logger is a singleton, configuration here will be used module-wide
+logger.remove()
+logger.add(
+    sys.stderr,
+    format="<green>{time:HH:mm:ss}</green> <bold>|</bold> <blue>{level}</blue> <bold>|</bold> {message}",
+    level="INFO",
+)
