+++ genomepy/providers/__init__.py
@@ -1,5 +1,5 @@
 import os
-from typing import Optional
+from typing import List, Optional
 
 import pandas as pd
 from loguru import logger
@@ -9,7 +9,7 @@ from genomepy.providers.ensembl import EnsemblProvider
 from genomepy.providers.ncbi import NcbiProvider
 from genomepy.providers.ucsc import UcscProvider
 from genomepy.providers.url import UrlProvider
-from genomepy.utils import best_search_result, get_genomes_dir, safe
+from genomepy.utils import get_genomes_dir, safe
 
 ASM_FORMAT = [
     "Sequence-Name",
@@ -48,16 +48,22 @@ def create(name: str):
     """
     name = name.lower()
     if name not in PROVIDERS:
-        logger.error(f"Unknown provider. Options: {', '.join(list_providers())}")
-        return
+        options = "', '".join(list_providers())
+        raise ValueError(f"Unknown provider '{name}'. Options: '{options}'")
 
-    p = PROVIDERS[name]()
-    return p
+    p = PROVIDERS[name]
+    p.download_assembly_report = staticmethod(download_assembly_report)
+    return p()
 
 
 def list_providers():
     """List available providers."""
-    return list(PROVIDERS.keys())
+    return [p.name for p in PROVIDERS.values()]
+
+
+def list_online_providers():
+    """Return a list of all providers that can be pinged."""
+    return [p.name for p in PROVIDERS.values() if p.ping()]
 
 
 def online_providers(provider: str = None):
@@ -72,7 +78,7 @@ def online_providers(provider: str = None):
             logger.warning(str(e))
 
 
-def search_all(term, provider: str = None, encode: bool = False):
+def search(term, provider: str = None):
     """
     Search for a genome.
 
@@ -86,36 +92,77 @@ def search_all(term, provider: str = None, encode: bool = False):
         Search term, case-insensitive.
     provider : str , optional
         Provider name
-    encode : bool, optional
-        Encode return strings.
 
     Yields
     ------
-    tuple
+    list
         genome information (name/identifier and description)
     """
     term = safe(str(term))
     for p in online_providers(provider):
         for row in p.search(term):
             ret = list(row[:1]) + [p.name] + list(row[1:])
-            if encode:
-                ret = [x.encode("utf-8") for x in ret]
             yield ret
 
 
-def nearest_ncbi_assembly(asm_acc):
+def _best_accession(reference: str, targets: list):
+    """Return the nearest accession ID from a list of IDs"""
+    if len(targets) == 1:
+        return targets[0]
+
+    # GCA/GCF
+    matching_prefix = [t for t in targets if t.startswith(reference[0:3])]
+    if len(matching_prefix) > 0:
+        targets = matching_prefix
+
+    # patch levels
+    # e.g. GCA_000002035.4 & GCA_000002035.3
+    ref_patch = int(reference.split(".")[1]) if "." in reference else 0
+    nearest = [999, []]
+    for target in targets:
+        tgt_patch = int(target.split(".")[1]) if "." in target else 0
+        distance = abs(ref_patch - tgt_patch)
+        if distance == nearest[0]:
+            nearest[1].append(target)
+        if distance < nearest[0]:
+            nearest = [distance, [target]]
+    targets = nearest[1]
+
+    if len(targets) > 1:
+        logger.info(
+            f"Multiple matching accession numbers found. Selecting the closest ({targets[0]})."
+        )
+    return targets[0]
+
+
+def _best_search_result(asm_acc: str, results: List[list]) -> list:
     """
-    return the NCBI accession ID and NCBI assembly name nearest to
-    the given accession ID.
+    Return the best search result based on accession IDs.
     """
-    ncbi_search = list(NcbiProvider().search(asm_acc))
-    search_result = best_search_result(asm_acc, ncbi_search, acc_idx=1)
-    if len(search_result) == 0:
-        logger.warning(f"Could not download an NCBI assembly report for {asm_acc}")
-        return None, None
-    ncbi_acc = search_result[1]
-    ncbi_name = search_result[0]
-    return ncbi_acc, ncbi_name
+    results = [res for res in results if res[2] is not None]
+
+    if len(results) > 1:
+        accessions = [res[2] for res in results]
+        bes_acc = _best_accession(asm_acc, accessions)
+        results = [res for res in results if res[2] == bes_acc]
+
+    if len(results) > 0:
+        return results[0]
+
+
+def nearest_assembly(asm_acc: str, provider: str) -> list:
+    """
+    Return the search result of the assembly nearest to
+    the given accession ID, in the specified provider.
+    """
+    if not asm_acc.startswith(("GCA_", "GCF_")):
+        raise ValueError("asm_acc must be an assembly accession ID.")
+
+    search_results = list(search(asm_acc, provider=provider))
+    best_result = _best_search_result(asm_acc, search_results)
+    if best_result is None:
+        logger.warning(f"No assembly similar to {asm_acc} on {provider}.")
+    return best_result
 
 
 def download_assembly_report(asm_acc: str, fname: str = None):
@@ -136,9 +183,11 @@ def download_assembly_report(asm_acc: str, fname: str = None):
     pandas.DataFrame
         NCBI assembly report.
     """
-    ncbi_acc, ncbi_name = nearest_ncbi_assembly(asm_acc)
-    if ncbi_acc is None:
+    search_result = nearest_assembly(asm_acc, "NCBI")
+    if search_result is None:
         return
+    ncbi_acc = search_result[2]
+    ncbi_name = search_result[0]
 
     # NCBI FTP location of assembly report
     assembly_report = (
@@ -196,7 +245,7 @@ def map_locations(
     if not os.path.exists(frm_asm_report):
         download_assembly_report(asm_acc, frm_asm_report)
     if not os.path.exists(frm_asm_report):
-        logger.error("Cannot map without an assembly report")
+        logger.warning("Cannot map without an assembly report.")
         return
 
     asm_report = pd.read_csv(frm_asm_report, sep="\t", comment="#")
@@ -214,9 +263,15 @@ def map_locations(
     if "ucsc" in [frm_provider, to_provider] and list(
         asm_report["ucsc_name"].unique()
     ) == ["na"]:
-        logger.error("UCSC style names not available for this assembly")
+        logger.warning("UCSC style names not available for this assembly.")
         return
 
     mapping = asm_report[[f"{frm_provider}_name", f"{to_provider}_name"]]
     mapping = mapping.dropna().drop_duplicates().set_index(f"{frm_provider}_name")
     return mapping
+
+
+class Provider:
+    list = staticmethod(list_providers)
+    online = staticmethod(list_online_providers)
+    create = staticmethod(create)
